optimized class structure add test cases
