+++ Tests/test_interface.py
@@ -1,0 +1,6 @@
+from AccountConsoleInterface import *
+
+dep1 = DepositAccount("125", 0.10, 0)
+dep2 = DepositAccount("126", 0.10, 0)
+
+AccountConsoleInterface.print_menu()
